moved Board.cs to models folder(forgot to do it earlier.
